VDraw: m_MsgTime type changes.
m_MsgTime is now an unsigned int, which matches GetTickCount().

writeText() now takes an unsigned short as the duration value.
I don't think there's any function that writes text to the screen
with a duration higher than 65,535 ms, anyway.

+++ src/gens/video/v_draw.cpp
@@ -553,7 +553,7 @@ void VDraw::setBpp(int newBpp)
  * @param msg Message to write.
  * @param duration Duration for the message to appear, in milliseconds.
  */
-void VDraw::writeText(string msg, int duration)
+void VDraw::writeText(string msg, unsigned short duration)
 {
 	if (!m_MsgEnabled)
 		return;
